+++ test/base/handlebarsTemplateEngine.test.js
@@ -22,7 +22,8 @@ config.templateEngine = {
 var suite = tools.createVowsSuite("Base: templateEngine/handlebarsTemplateEngine", {
   config: config,
   applications: null,
-  useRedis: false
+  useRedisSocketStore: false,
+  useRedisSessionStore: false
 });
 tools.addBatches(suite, "handlebarsTemplateEngine", "general");