+++ test/cluster/startup.test.js
@@ -14,7 +14,8 @@ var config = clone(clusterConfig);
 var suite = tools.createVowsSuiteForCluster("Cluster: Startup Thywill", {
   config: config,
   applications: null,
-  useRedis: true
+  useRedisSocketStore: true,
+  useRedisSessionStore: true
 });
 
 //-----------------------------------------------------------