+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Work in Progress
 ----------------
 
   * Expanded test coverage to applications with web/socket client tests.
+  * Crucial bugfix for sessions in SocketIoExpressClientInterface.
 
 v0.0.8 - 04/06/2013
 -------------------